+++ src/main/java/com/doc/demo/stream/impl/LocalDocStreamImpl.java
@@ -99,7 +99,7 @@ public class LocalDocStreamImpl implements DocStream<DocLocalParam, DocLocalResu
     }
 
     @Override
-    public String docPreview(@NotNull DocLocalParam param, @NotNull Long expiredTime) {
+    public String docPreview(@NotNull DocLocalParam param, @NotNull Integer expiredTime) {
         byte[] bytes = this.downloadDoc(param);
         File previewFile = getPreviewFileDir(param);
         //获取预览的根路径文件
